功能基本ok
基本功能ok,待优化逻辑、布局
1、功能是否完整、是否存在BUG
2、布局调整

+++ WifiBoardDetect V1.0.0/src/com/zws/wifiboarddetect/MyListItem.java
@@ -2,9 +2,9 @@ package com.zws.wifiboarddetect;
 
 public class MyListItem {
 	String ssid;
-	int level;
+	float level;
 
-	public MyListItem(String pssid, int plevel) {
+	public MyListItem(String pssid, float plevel) {
 		ssid = pssid;
 		level = plevel;
 	}
